Redesign of the watch webpage
The plan is to get rid of the classless.css file and replace it with more lightweight stylesheet that will allow more flexible styling.
The video now fills as much of the screen as possible.
Subtitles now have dark grey background instead of the black one they used to have.