Add a butt ton of libraries to export into Anki packages
And it works! Two templates (card styles) ara available at the moment) and the form for exporting will be worked on in the next few commits.
New carousel component was added, that is coded in separate CSS and JS files, as it might be used on other webpages too later.